requirements.txt
Django~=5.2.1
djangorestframework~=3.16.0
djangorestframework-simplejwt
django-cors-headers
psycopg2
django-filter
